+++ src/utils/crud.tsx
@@ -1,5 +1,5 @@
 const LOCAL_SERVER = false;
-const SERVER_URL = "https://home-library-server.herokuapp.com";
+const SERVER_URL = "https://kitaabghar-server.herokuapp.com";
 const LOCAL_SERVER_URL = "192.168.1.67";
 const LOCAL_SERVER_PORT = 4000;
 
@@ -12,24 +12,6 @@ function getServerUrl() {
 }
 
 export const SERVER = getServerUrl();
-
-export enum Status {
-  ERROR = "error",
-  SUCCESS = "success",
-}
-
-export enum Method {
-  PUT = "PUT",
-  GET = "GET",
-  POST = "POST",
-  DELETE = "DELETE",
-}
-
-export enum viewState {
-  MODULE = "module",
-  LIST = "list",
-  HEADLINE = "headline",
-}
 export const EXPORT_FILE_EXTENSION = ".csv";
 export const MAXIMUM_VISIBLE_DESCRIPTION_LENGTH = 500;
